Match scroll_flag type with fmm_disp.

+++ trunk/src/nh99/admodel.h
@@ -1830,7 +1830,7 @@ public:
   long iexit;
   long ihflag;
   long ihang;
-  long scroll_flag;
+  int scroll_flag;
   int maxfn_flag;
   int quit_flag;
   double min_improve;
